backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.5
fastapi==0.135.1
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
greenlet==3.3.2
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.31.2
idna==3.11
iniconfig==2.3.0
mako==1.3.10
markupsafe==3.0.3
overkill==0.4.0
packaging==26.0
playwright==1.58.0
pluggy==1.6.0
proto-plus==1.27.1
protobuf==5.29.6
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.13.1
pyee==13.0.1
pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
python-dotenv==1.2.2
requests==2.32.5
soupsieve==2.8.3
sqlalchemy==2.0.48
starlette==0.52.1
tqdm==4.67.3
typing-extensions==4.15.0
typing-inspection==0.4.2
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.42.0
yt-dlp==2026.3.13
